MGMT-4872: Initial events generation
Events generation is designed to expose a uniformed method of
initializing the assisted-service events, based on each event definition.
The event definition will require a set of parameters for each event type.
It will be used for generating a function for initializing the event with
the provided parameters, hence enforcing to provide the parameters at
initialization and not to miss any needed information.
The events will be defined in `./data/events.yaml`.

In order to introduce a new event, the developer will have to follow the
next steps:
1. Add event definition to `data/events.yaml` (event definition will be
   discussed ahead).
2. Generate the code for creating the event by: `skipper make generate
   generate-events`
3. Use the generated function for the event from
   `./internal/common/events.event.go` for creating the event to be
   emitted by the event handler.

Event definition should specify the following attributes:
- name: The name of the event which needs to remain unique and constant
  as it may be referred by the service's clients (e.g. by the UI).
- message: A template of the message that will be rendered if it
  contains references to the properties. E.g. the message "Install
  cluster {cluster_id}" expects the existence of a property named
  cluster_id
- event_type: Can be either "cluster" or "host". "cluster" type requires
  the existence of "cluster_id" in properties and "host" type requires
  the existence of "host_id".
- severity: Any of info, warning, error or critical
- properties: A list of properties to be rendered into the message (if
  referred by) or added as identifier of the message (e.g. cluster_id, host_id)
  with the type of the property.